requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
dj-database-url==0.5.0
Django==2.1.7
django-heroku==0.3.1
djangorestframework==3.9.2
djangorestframework-stubs==0.1.0
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gspread==3.1.0
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
numpy==1.16.2
oauth2client==4.1.3
oauthlib==3.0.1
panda==0.3.1
pandas==0.24.1
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyDrive==1.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
whitenoise==4.1.2
